requirements.txt
neat-python==0.92
numpy==2.0.1
opencv-python==4.10.0.84
gym-retro